pyproject.toml
[tool.poetry]
name = "oncallloader"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.4.1"
pydantic-settings = "^2.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
punq = "^0.6.2"
pydantic-yaml = "^1.2.0"
python-dateutil = "^2.8.2"
prometheus-client = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
